Add chart file
